+++ src/Plainion.GraphViz.Infrastructure/ViewModel/DefaultDomainModel.cs
@@ -9,18 +9,11 @@ namespace Plainion.GraphViz.Infrastructure.ViewModel
     public class DefaultDomainModel : BindableBase, IDomainModel
     {
         private IGraphPresentation myPresentation;
-        private ILayoutEngine myLayoutEngine;
 
         public IGraphPresentation Presentation
         {
             get { return myPresentation; }
             set { SetProperty( ref myPresentation, value ); }
-        }
-
-        public ILayoutEngine LayoutEngine
-        {
-            get { return myLayoutEngine; }
-            set { SetProperty( ref myLayoutEngine, value ); }
         }
     }
 }